QOL & Bugfixes
Sqlite database no longer throws an exception when it's first being initialised.
MediaView now waits for the pages to be rendered before rendering the main issue (solving a bug where we didn't know the maximum number of possible pages when checking for whether the current page ID went over it).
Clicking on the main image in the image view now opens just the image in a new tab
Page buttons now try to transition between index pages instead of only displaying the buttons for 1 page.